app/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
requests==2.32.3
octodns==1.15.0
pyyaml==6.0.3
cryptography==44.0.3
